pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "SEEM"
version = "1.0"
description = "Segment Everything Everywhere All at Once."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
]

dependencies = [
    "torch",
    "torchvision",
    "pillow==9.4.0",
    "opencv-python==4.8.1.78",
    "pyyaml==6.0.1",
    "json_tricks==3.17.3",
    "yacs==0.1.8",
    "scikit-learn==1.3.1",
    "pandas==2.0.3",
    "timm==0.4.12",
    "numpy==1.23.1",
    "einops==0.7.0",
    "fvcore==0.1.5.post20221221",
    "transformers==4.34.0",
    "sentencepiece==0.1.99",
    "ftfy==6.1.1",
    "regex==2023.10.3",
    "nltk==3.8.1",
    "vision-datasets==0.2.2",
    "cython==3.0.2",
    "pycocotools==2.0.7",
    "diffdist==0.1",
    "pyarrow==13.0.0",
    "cityscapesscripts==2.2.2",
    "shapely==1.8.0",
    "scikit-image==0.21.0",
    "mup==1.0.0",
    "accelerate==0.23.0",
    "kornia==0.6.8",
    "deepspeed==0.10.3",
    "wandb==0.15.12",
    "infinibatch==0.1.1",
    "gradio==3.42.0",     
    "openai-whisper",    
]

[tool.poetry.dependencies]
detectron2 = {git = "https://github.com/MaureenZOU/detectron2-xyz.git"}


[project.urls]
"Paper" = "https://arxiv.org/abs/2304.06718"
"Code" = "https://github.com/UX-Decoder/Segment-Everything-Everywhere-All-At-Once/tree/v1.0"
"Bug Tracker" = "https://github.com/UX-Decoder/Segment-Everything-Everywhere-All-At-Once/issues"

[tool.setuptools.packages.find]
exclude = ["assets*", "demo*"]

[tool.wheel]
exclude = ["assets*", "demo*"]
